# Conflicts:
#	dataprep-api/src/main/java/org/talend/dataprep/api/service/settings/context/provider/ContextProvider.java
#	dataprep-backend-service/src/main/resources/org/talend/dataprep/configuration/default.properties
#	dataprep-webapp/config/sass.conf.js
#	dataprep-webapp/config/webpack.config.js
#	dataprep-webapp/package.json
#	dataprep-webapp/src/app/index-module.js
#	dataprep-webapp/src/index.html
#	dataprep-webapp/yarn.lock

+++ dataprep-preparation/src/main/java/org/talend/dataprep/preparation/service/FolderService.java
@@ -13,7 +13,9 @@
 package org.talend.dataprep.preparation.service;
 
 import static java.util.stream.Collectors.toList;
-import static org.springframework.web.bind.annotation.RequestMethod.*;
+import static org.springframework.web.bind.annotation.RequestMethod.DELETE;
+import static org.springframework.web.bind.annotation.RequestMethod.GET;
+import static org.springframework.web.bind.annotation.RequestMethod.PUT;
 import static org.talend.daikon.exception.ExceptionContext.build;
 import static org.talend.dataprep.api.folder.FolderContentType.PREPARATION;
 import static org.talend.dataprep.exception.error.FolderErrorCodes.FOLDER_NOT_FOUND;
@@ -27,13 +29,19 @@ import java.util.stream.StreamSupport;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 import org.talend.daikon.exception.ExceptionContext;
 import org.talend.dataprep.api.folder.Folder;
 import org.talend.dataprep.api.folder.FolderInfo;
 import org.talend.dataprep.api.folder.FolderTreeNode;
 import org.talend.dataprep.exception.TDPException;
 import org.talend.dataprep.folder.store.FolderRepository;
+import org.talend.dataprep.http.HttpResponseContext;
 import org.talend.dataprep.metrics.Timed;
 import org.talend.dataprep.security.Security;
 import org.talend.dataprep.util.SortAndOrderHelper.Order;
@@ -174,7 +182,11 @@ public class FolderService {
     @ApiOperation(value = "Remove a Folder", notes = "Remove the folder")
     @Timed
     public void removeFolder(@PathVariable String id) {
-        folderRepository.removeFolder(id);
+        if (!folderRepository.exists(id)) {
+            HttpResponseContext.status(HttpStatus.NOT_FOUND);
+        } else {
+            folderRepository.removeFolder(id);
+        }
     }
 
     /**
